requirements/common.txt
# Core Stuff
# -------------------------------------
Django==3.2.11
six==1.16.0

# Configuration
# -------------------------------------
argon2-cffi==21.3.0
django-environ==0.8.1
django-sites==0.11
python-dotenv==0.19.2
django-cors-headers==3.11.0
django-solo==2.0.0
phonenumbers==8.12.42
django-phonenumber-field==6.0.0
#warrant==0.6.1
pycognito
createsend==6.1.2  # for campaign monitor
django-regex-field==2.0.2
ua-parser==0.10.0
user-agents==2.2.0
django-user-agents==0.4.0

# Extensions
# -------------------------------------
pytz==2021.3
pyzipcode==3.0.1
timezonefinder==5.2.0

# Models
# -------------------------------------
psycopg2-binary==2.9.3
django-reversion==4.0.2
Pillow==9.0.1
Pygments==2.11.2
django-extensions==3.1.5
django-uuid-upload-path==1.0.0
django-versatileimagefield==2.2
django-extended-choices==1.3.3

# REST APIs
# -------------------------------------
djangorestframework==3.13.1
django-rest-swagger==2.2.0

# LOGGING
# -------------------------------------
django-log-request-id==2.0.0

# Documentation
# -------------------------------------
mkdocs==1.3.0
jinja2==3.0.3

# Raven is the Sentry client
# --------------------------
raven==6.10.0

# Async Tasks
# -------------------------------------
celery[redis]==5.2.2
flower==1.0.0

# Auth Stuff
# -------------------------------------
PyJWT==2.3.0
django-mail-templated==2.6.5

# Utils
# -------------------------------------
requests==2.27.1
pandas==1.4.0
django-admin-sortable2==1.0.3
dictdiffer==0.9.0  # https://dictdiffer.readthedocs.io/en/latest/
### added new library
django-debug-toolbar
#### Api key
djangorestframework-api-key==2.*
